move and update getPathWithoutQueryOrHash -> splitPathFromQueryAndFragment (add tests)
